chore: consolidate util functions
